pyproject.toml
[project]
name = "navi-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "KangJiUng",email = "woong236406@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "ulid-py (>=1.1.0,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

